Update build config

+++ config/openresty/build-config.js
@@ -21,9 +21,13 @@ const partials = {};
 const config_directory =
   process.env.OPENRESTY_CONFIG_DIRECTORY || "/home/ec2-user/openresty";
 
-
-// strip the trailing 'B' from 'MB' in config.webhooks.client_max_body_size
-const webhooks_client_max_body_size = config.webhooks.client_max_body_size.replace(/b/i, "");
+// max file size for webhooks bodies
+// nginx requires 'M' instead of 'MB' but unfortunately
+// node rawbody parser requires 'MB' instead of 'M'
+// so this maps '25MB' to '25M' for nginx
+const webhooks_client_max_body_size = config.webhooks.client_max_body_size
+  ? config.webhooks.client_max_body_size.replace(/MB/i, "M")
+  : "100M";
 
 const locals = {
   host: "blot.im",
@@ -58,19 +62,19 @@ const locals = {
   ssl_certificate:
     process.env.SSL_CERTIFICATE || "/etc/ssl/private/letsencrypt-domain.pem",
   ssl_certificate_key:
-    process.env.SSL_CERTIFICATE_KEY || "/etc/ssl/private/letsencrypt-domain.key"
+    process.env.SSL_CERTIFICATE_KEY ||
+    "/etc/ssl/private/letsencrypt-domain.key",
 };
 
 // move the previous contents of the data directory to a backup
 // so we can compare the new contents with the old
-if (fs.existsSync(OUTPUT))
-  fs.moveSync(OUTPUT, PREVIOUS, { overwrite: true });
+if (fs.existsSync(OUTPUT)) fs.moveSync(OUTPUT, PREVIOUS, { overwrite: true });
 
 fs.emptyDirSync(OUTPUT);
 
 fs.copySync(`${__dirname}/html`, `${OUTPUT}/html`);
 
-fs.readdirSync(CONFIG_DIRECTORY).forEach(file => {
+fs.readdirSync(CONFIG_DIRECTORY).forEach((file) => {
   // copy lua files to data directory so they are available to nginx
   if (file.endsWith(".lua")) {
     fs.copySync(CONFIG_DIRECTORY + "/" + file, OUTPUT + "/" + file);
@@ -109,7 +113,11 @@ if (process.argv.includes("--skip-confirmation")) {
 }
 
 // compare the new contents with the old
-const diff = child_process.spawnSync("/opt/homebrew/bin/diff", ['--color', "-r", PREVIOUS, OUTPUT ], { stdio: 'inherit' });
+const diff = child_process.spawnSync(
+  "/opt/homebrew/bin/diff",
+  ["--color", "-r", PREVIOUS, OUTPUT],
+  { stdio: "inherit" }
+);
 
 if (diff.error) {
   console.error(diff.error);
@@ -124,11 +132,11 @@ if (diff.error) {
 
   const rl = readline.createInterface({
     input: process.stdin,
-    output: process.stdout
+    output: process.stdout,
   });
 
   const question = () => {
-    rl.question("Do you want to keep these changes? [y/n] ", answer => {
+    rl.question("Do you want to keep these changes? [y/n] ", (answer) => {
       if (answer === "y") {
         console.log("Changes kept. Build complete");
         rl.close();
@@ -150,4 +158,3 @@ if (diff.error) {
 
   question();
 }
-
